requirements.txt
python-consul==1.1.0
flask==1.0.2